Separate test fixture in common.py file
